Create FSS trait
