+++ funcs/src/point/tests.rs
@@ -56,16 +56,17 @@ fn test_correctness_helper<D: DPF<F>>() {
         let y = F::rand(&mut rng);
 
         // Create the DPF
-        let (key1, key2) = D::gen(log_domain, x, y, &mut rng).unwrap();
+        let func = (log_domain, x, y);
+        let (key1, key2) = D::gen(&func, &mut rng).unwrap();
 
         // Evaluate each point of the DPF
         for p in valid_range {
-            let p1_result = D::eval(&key1, p).unwrap();
-            let p2_result = D::eval(&key2, p).unwrap();
+            let p1_result = D::eval(&key1, &p).unwrap();
+            let p2_result = D::eval(&key2, &p).unwrap();
             if p == x {
-                assert!(p1_result - p2_result == y)
+                assert!(D::decode((&p1_result, &p2_result)).unwrap() == y)
             } else {
-                assert!(p1_result - p2_result == F::zero())
+                assert!(D::decode((&p1_result, &p2_result)).unwrap() == F::zero())
             }
         }
     }
@@ -79,18 +80,20 @@ fn test_bad_inputs_helper<D: DPF<F>>() {
 
     // Test Gen fail
     let x = rng.gen_range(bad_range.clone());
-    let result = D::gen(log_domain, x, F::one(), &mut rng);
+    let func = (log_domain, x, F::one());
+    let result = D::gen(&func, &mut rng);
     assert!(result.is_err());
 
     // Test P1 Eval fail
     let x = rng.gen_range(0..max);
-    let (k1, k2) = D::gen(log_domain, x, F::one(), &mut rng).unwrap();
+    let func = (log_domain, x, F::one());
+    let (k1, k2) = D::gen(&func, &mut rng).unwrap();
 
     let p = rng.gen_range(bad_range);
-    let result = D::eval(&k1, p);
+    let result = D::eval(&k1, &p);
     assert!(result.is_err());
 
-    let result = D::eval(&k2, p);
+    let result = D::eval(&k2, &p);
     assert!(result.is_err());
 }
